Rename mocked balance states to exchangeBalanceStates
