+++ test/data/index.ts
@@ -1,4 +1,4 @@
-export { default as balanceStates } from './balanceStates'
+export { default as exchangeBalanceStates } from './exchangeBalanceStates'
 
 export const CONTRACT = '0x1000000000000000000000000000000000000000'
 export const ADDRESS = '0x1111111111111111111111111111111111111111'
